+++ doc/whats_new.rst
@@ -11,6 +11,8 @@ Bug fixes
 
 - Fix misinterpreted tuples passed as ``allowed_dims`` argument of
   ``Variable`` init (:issue:`17`).
+- Better error message when a Model instance is expected but no object
+  is found or a different object is provided (:issue:`13`).
 
 v0.1.0 (8 October 2017)
 -----------------------